pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "strands-agents-evals"
dynamic = ["version"]
description = "Evaluation framework for Strands"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "AWS", email = "opensource@amazon.com"},
]

dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "rich>=14.0.0,<15.0.0",
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.1.0,<1.0.0",
    "typing-extensions>=4.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/strands_evals"]

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=4.0",
]

dev = [
    "hatch>=1.0.0,<2.0.0",
    "mypy>=1.0",
    "pre-commit>=3.2.0,<4.2.0",
    "ruff>=0.4.4,<1.0.0",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
follow_untyped_imports = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "strands_evals.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 120
include = ["src/**/*.py", "tests/**/*.py"]

[tool.hatch.envs.default.scripts]
list = [
    "echo 'Scripts commands available for default env:'; hatch env show --json | jq --raw-output '.default.scripts | keys[]'"
]
format = [
  "hatch fmt --formatter",
]
test-format = [
    "hatch fmt --formatter --check",
]
lint = [
    "hatch fmt --linter"
]
test-lint = [
    "hatch fmt --linter --check"
]
prepare = [
    "hatch fmt --linter",
    "hatch fmt --formatter",
    "hatch run test-lint",
    "hatch test --all"
]


[tool.hatch.envs.hatch-static-analysis.scripts]
format-check = [
    "ruff format --check"
]
format-fix = [
    "ruff format"
]
lint-check = [
    "ruff check",
    "mypy -p src"
]
lint-fix = [
    "ruff check --fix"
]

[tool.ruff.lint]
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I", # isort
  "B", # flake8-bugbear
]

[tool.hatch.version]
path = "src/strands_evals/__init__.py"
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.0",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=4.0",
]
extra-dependencies = [
    "hatch>=1.0.0,<2.0.0",
    "mypy>=1.0",
    "pre-commit>=3.2.0,<4.2.0",
    "ruff>=0.4.4,<1.0.0",
]